Use column names in format string
This replaces the, somwhat arbitrary, names that were chosen for the
format string with column names in all cases but intelligibility.
Inteligibility is still kept as one variable  to make it easier to pass
to format. This makes reprocess easier and more robust because it
doesn't have to know the format string names or all column names.